add command to change user's admin and activity status
remove permission backend setup from auth and move to own app (auth_permissions.NewAppConfig)

Changes in comments in storage_test for audit_logs_mysql

+++ src/contrib/pages/backend-sqlite/schema.sqlite3.sql
@@ -18,16 +18,16 @@ CREATE TABLE IF NOT EXISTS PageNode (
     -- 0x01: Published
     -- 0x02: Hidden
     -- 0x04: Deleted
-    status_flags    BIGINT    NOT NULL,
+    status_flags    INTEGER    NOT NULL,
 
     -- The page ID that this node represents
-    page_id         BIGINT    NOT NULL,
+    page_id         INTEGER    NOT NULL,
 
     -- The unique content type name for this node
     content_type    TEXT      NOT NULL,
 
     -- Latest revision ID
-    latest_revision_id BIGINT,
+    latest_revision_id INTEGER,
 
     created_at      TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP, -- Read-only
     updated_at      TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP -- Read-only
